Refactored assignments into separate component.
