+++ client/src/parent/ChoreTableRow.tsx
@@ -1,21 +1,10 @@
-import {
-  TableRow,
-  TableCell,
-  Checkbox,
-  Button,
-  makeStyles,
-} from "@material-ui/core";
+import { TableRow, TableCell, makeStyles } from "@material-ui/core";
+import { ChoreAssignment } from "./ChoreAssignment";
 import { ChoreDetails } from "./ChoreDetails";
-import { AssignmentState } from "./models/AssignmentState";
 import { ChoreWithAssignments } from "./models/ChoreWithAssignments";
 
 const useStyles = makeStyles({
   row: {
-    "& > *": {
-      borderBottom: "unset",
-    },
-  },
-  assignmentRow: {
     "& > *": {
       borderBottom: "unset",
     },
@@ -29,23 +18,6 @@ type Prop = {
 export function ChoreTableRow(props: Prop) {
   const classes = useStyles();
 
-  function isDone(state: AssignmentState): boolean {
-    return (
-      state === AssignmentState.RequestedToCheck ||
-      state === AssignmentState.CheckConfirmed ||
-      state === AssignmentState.CheckRefused ||
-      state === AssignmentState.Archived
-    );
-  }
-
-  function isConfirmed(state: AssignmentState): boolean {
-    return (
-      state === AssignmentState.CheckConfirmed ||
-      state === AssignmentState.CheckRefused ||
-      state === AssignmentState.Archived
-    );
-  }
-
   return (
     <>
       <TableRow className={classes.row} key={props.chore.id}>
@@ -55,33 +27,10 @@ export function ChoreTableRow(props: Prop) {
       </TableRow>
       {props.chore.assignments.map((assignment, index) => {
         return (
-          <TableRow
-            className={
-              index !== props.chore.assignments.length - 1
-                ? classes.assignmentRow
-                : ""
-            }
-            key={`${props.chore.id}-${assignment.userId}`}
-          >
-            <TableCell component="th" scope="row"></TableCell>
-            <TableCell align="left">{assignment.userName}</TableCell>
-            <TableCell align="center">
-              <Checkbox
-                checked={isDone(assignment.state)}
-                disabled
-                size="small"
-              />
-            </TableCell>
-            <TableCell align="center">
-              {isDone(assignment.state) && !isConfirmed(assignment.state) ? (
-                <Button variant="contained" color="primary" size="small">
-                  Ok
-                </Button>
-              ) : (
-                ""
-              )}
-            </TableCell>
-          </TableRow>
+          <ChoreAssignment
+            assignment={assignment}
+            isLastRow={index === props.chore.assignments.length - 1}
+          />
         );
       })}
     </>
